requirements.txt
jupyter==1.1.1
pandas==2.2.3
scikit-learn==1.6.1
boto3==1.38.7
numpy==2.2.5
scipy==1.15.2
matplotlib==3.10.1
joblib == 1.5.0
mlflow==2.22.0
awscli==1.40.22
flask==3.1.1
flask_cors==6.0.0
python-dotenv==1.1.0